Refactored code, added a multiprocess inverter
