+++ page_ranker_app/__init__.py
@@ -1,0 +1,5 @@
+import sys
+import os
+
+
+sys.path.append(os.path.dirname(__file__))
